requirements.txt
asgiref==3.6.0
Django==4.1.7
django-jazzmin==2.6.0
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
